Deduplicate setup_logging test utility
