+++ p2panda-auth/src/test_utils/mod.rs
@@ -181,11 +181,3 @@ pub fn assert_members(
     expected.sort();
     assert_eq!(actual, expected);
 }
-
-pub fn setup_logging() {
-    if std::env::var("RUST_LOG").is_ok() {
-        let _ = tracing_subscriber::fmt()
-            .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
-            .try_init();
-    }
-}
