+++ p2panda-net/src/test_utils.rs
@@ -139,14 +139,6 @@ pub fn test_args_from_seed(seed: [u8; 32]) -> ApplicationArguments {
         .build()
 }
 
-pub fn setup_logging() {
-    if std::env::var("RUST_LOG").is_ok() {
-        let _ = tracing_subscriber::fmt()
-            .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
-            .try_init();
-    }
-}
-
 #[test]
 fn deterministic_args() {
     let args_1 = test_args_from_seed([0; 32]);
