+++ p2panda-sync/src/test_utils.rs
@@ -145,14 +145,6 @@ impl Peer {
     }
 }
 
-pub fn setup_logging() {
-    if std::env::var("RUST_LOG").is_ok() {
-        let _ = tracing_subscriber::fmt()
-            .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
-            .try_init();
-    }
-}
-
 /// Run a pair of topic sync sessions.
 pub async fn run_protocol<P>(
     session_local: P,
